requirements.txt
amqp (1.4.6)
anyjson (0.3.3)
billiard (3.3.0.19)
celery (3.1.17)
Django (1.7.4)
docker-py (0.7.2)
gitdb (0.6.4)
GitPython (0.3.6)
kombu (3.0.24)
pip (1.4.1)
pytz (2014.10)
requests (2.5.1)
setuptools (0.9.8)
six (1.9.0)
smmap (0.9.0)
websocket-client (0.23.0)
wsgiref (0.1.2)
